# Conflicts:
#	autobuild.xml - Removed abseil
#	indra/cmake/Abseil-CPP.cmake - moved to deps/CMakeLists.txt
#	indra/cmake/CMakeLists.txt
#	indra/cmake/Linking.cmake
#	indra/cmake/Variables.cmake - merge with alch
#	indra/llcommon/indra_constants.* - merge
#	indra/llcommon/llmemory.* - merge
#	indra/llcommon/llsys.* - merge
#	indra/llmath/llmatrix3a.h
#	indra/llmath/llmatrix4a.h
#	indra/llmath/llquaternion2.h
#	indra/llmath/llsimdtypes.h
#	indra/llmath/llvector4a.h
#	indra/llplugin/llpluginclassmedia.cpp
#	indra/llui/CMakeLists.txt
#	indra/newview/CMakeLists.txt
#	indra/newview/llappviewer.cpp
#	indra/newview/llappviewerwin32.cpp
#	indra/newview/llfloaterland.cpp
#	indra/newview/llfloaterland.h
#	indra/newview/llmaterialmgr.h
#	indra/newview/llmeshrepository.h
#	indra/newview/lltexturefetch.cpp - shy's texture fetch change ported
#	indra/newview/llviewerdisplay.cpp
#	indra/newview/llviewerinventory.cpp
#	indra/newview/llviewermenu.cpp
#	indra/newview/llviewerobject.cpp
#	indra/newview/llviewerregion.h
#	indra/newview/llviewertexture.cpp
#	indra/newview/llviewertexture.h
#	indra/newview/llviewertexturelist.cpp
#	indra/newview/llviewertexturelist.h
#	indra/newview/llviewerwearable.cpp
#	indra/newview/llvoavatar.cpp
#	indra/newview/llvoavatarself.h
#	indra/newview/llvovolume.h
#	indra/newview/viewer_manifest.py

+++ indra/newview/llpreviewscript.cpp
@@ -720,7 +720,7 @@ bool LLScriptEdCore::loadScriptText(const std::string& filename)
 
 	// read in the whole file
 	fseek(file, 0L, SEEK_END);
-	long file_length = ftell(file);
+	size_t file_length = ftell(file);
 	fseek(file, 0L, SEEK_SET);
 	if (file_length > 0)
 	{
